Add files via upload
Prepare a model for car price prediction with the available variables. To help the management decide significant variables in predicting the price of a car. The data used here has been gathered by a survey.
The model will be a good way for management to understand the pricing dynamics of a new market.